Rename bold_table_lint.py to iso_bold_table_lint.py
Consistent naming with iso_heading_case_lint.py and iso_clause_lint.py.
